+++ src/event-listeners/KeyDownEventListener.test.ts
@@ -1,6 +1,6 @@
 import { beforeEach, describe, expect, it, SpyInstance, vitest } from 'vitest'
 import UserActionHandler from '../user-action/UserActionHandler'
-import { fireEvent, getAllByRole, getByRole, waitFor } from '@testing-library/dom'
+import { fireEvent, getByRole, getByTestId, waitFor } from '@testing-library/dom'
 import { nop } from '../utils/nop'
 import createElementInJSDOM from '../test-utils/createElementInJSDOM'
 import UserActionsHistory from '../user-actions-history/UserActionsHistory'
@@ -9,6 +9,7 @@ import MemorySessionIdHandler from '../session-id-handler/MemorySessionIdHandler
 import KeyDownEventListener from './KeyDownEventListener'
 import * as createTargetedUserActionModule from '../user-action/createTargetedUserAction'
 import ISessionIdHandler from '../session-id-handler/ISessionIdHandler'
+import { TargetedUserAction, UserActionType } from '../types'
 
 describe('KeyDownEventListener', () => {
   let userActionHistory: UserActionsHistory
@@ -26,7 +27,7 @@ describe('KeyDownEventListener', () => {
     createTargetedUserActionSpy = vitest.spyOn(createTargetedUserActionModule, 'createTargetedUserAction')
   })
 
-  it('it calls createTargetedUserAction with the excludeRegex option', async () => {
+  it('calls createTargetedUserAction with the excludeRegex option', async () => {
     const { element, domWindow } = createElementInJSDOM(
       `
             <div>
@@ -47,7 +48,7 @@ describe('KeyDownEventListener', () => {
     })
   })
 
-  it('it calls createTargetedUserAction with the customSelector option', async () => {
+  it('calls createTargetedUserAction with the customSelector option', async () => {
     const { element, domWindow } = createElementInJSDOM(
       `
             <div>
@@ -90,9 +91,9 @@ describe('KeyDownEventListener', () => {
   it('calls handler when key down event been fired on a div', async () => {
     const { element, domWindow } = createElementInJSDOM(
       `
-                <div>
-                    <div role='cell'/>
-                </div>`,
+        <div>
+            <div role='cell'/>
+        </div>`,
       'div',
     )
 
@@ -102,6 +103,7 @@ describe('KeyDownEventListener', () => {
     fireEvent.keyDown(div)
 
     await waitFor(() => {
+      console.log()
       expect(handleSpy).toHaveBeenCalledOnce()
     })
   })
@@ -109,16 +111,16 @@ describe('KeyDownEventListener', () => {
   it('does not call listener multiple time when key down event is fired consecutively', async () => {
     const { element, domWindow } = createElementInJSDOM(
       `
-                <div>
-                    <div role='cell'/>
-                </div>`,
+        <div>
+            <div role='cell'/>
+        </div>`,
       'div',
     )
 
     new KeyDownEventListener(userActionHandler, domWindow, userActionHistory).init()
     const div = await waitFor(() => getByRole(element, 'cell'))
 
-    for (let i = 0; i < 10; i++) fireEvent.keyDown(div)
+    for (let i = 0; i < 10; i++) fireEvent.keyDown(div, { code: 'Enter' })
 
     await waitFor(() => {}, { timeout: 200 })
 
@@ -127,71 +129,167 @@ describe('KeyDownEventListener', () => {
     })
   })
 
-  it('does not call listener when key down event is fired on text inputs', async () => {
-    const { element, domWindow } = createElementInJSDOM(
-      `
-            <div>
-                <input id='text-1' type='text'/>
-                <textarea id='text-2'></textarea>
-                <input id='text-5' type='search' />
-            </div>`,
-      'div',
-    )
+  describe('when a special key is typed', () => {
+    const inputs = [
+      {
+        query: (element: HTMLElement) => getByRole(element, 'textbox'),
+        inputType: 'text',
+        html: "<input id='text-1' type='text'/>",
+        expectedUserActionType: UserActionType.KeyDown,
+      },
+      {
+        query: (element: HTMLElement) => getByRole(element, 'textbox'),
+        inputType: 'textarea',
+        html: "<textarea id='text-2'></textarea>",
+        expectedUserActionType: UserActionType.KeyDown,
+      },
+      {
+        query: (element: HTMLElement) => getByRole(element, 'searchbox'),
+        inputType: 'search',
+        html: "<input id='text-5' type='search' />",
+        expectedUserActionType: UserActionType.KeyDown,
+      },
+    ]
 
-    new KeyDownEventListener(userActionHandler, domWindow, userActionHistory).init()
+    for (const { html, inputType, query, expectedUserActionType } of inputs) {
+      it(`calls handler when the key is Tab on ${inputType}`, async () => {
+        await assertHandleUserActionIsCalled('Tab', html, query, expectedUserActionType)
+      })
 
-    const inputs: HTMLElement[] = await waitFor(() => getAllByRole(element, 'textbox'))
-    const search = await waitFor(() => getByRole(element, 'searchbox'))
+      it(`calls handler when the key is Enter on ${inputType}`, async () => {
+        await assertHandleUserActionIsCalled('Enter', html, query, expectedUserActionType)
+      })
 
-    for (const input of inputs) {
-      fireEvent.keyDown(input)
+      it(`calls handler when the key is NumpadEnter on ${inputType}`, async () => {
+        await assertHandleUserActionIsCalled('NumpadEnter', html, query, expectedUserActionType)
+      })
     }
-    fireEvent.keyDown(search)
 
-    await waitFor(() => {}, { timeout: 500 })
-
-    await waitFor(() => {
-      expect(handleSpy).toHaveBeenCalledTimes(0)
-    })
-  })
-
-  describe('When press on allowed key', () => {
-    it('calls handler when the key is Tab', async () => {
-      await assertHandleUserActionIsCalled('Tab')
-    })
-
-    it('calls handler when the key is Enter', async () => {
-      await assertHandleUserActionIsCalled('Enter')
-    })
-
-    it('calls handler when the key is NumpadEnter', async () => {
-      await assertHandleUserActionIsCalled('NumpadEnter')
-    })
-
-    async function assertHandleUserActionIsCalled(code: string) {
-      const { element, domWindow } = createElementInJSDOM(
-        `
-            <div>
-                <input id='text-1' type='text'/>
-                <textarea id='text-2'></textarea>
-                <input id='text-5' type='search' />
-            </div>`,
-        'div',
-      )
+    async function assertHandleUserActionIsCalled(
+      code: string,
+      domContent: string,
+      query: (element: HTMLElement) => HTMLElement,
+      expectedUserActionType: UserActionType,
+    ) {
+      const { element, domWindow } = createElementInJSDOM(`<div>${domContent}</div>`, 'div')
 
       new KeyDownEventListener(userActionHandler, domWindow, userActionHistory).init()
 
-      const inputs: HTMLElement[] = await waitFor(() => getAllByRole(element, 'textbox'))
-      const search = await waitFor(() => getByRole(element, 'searchbox'))
-
-      for (const input of inputs) {
-        fireEvent.keyDown(input, { code })
-      }
-      fireEvent.keyDown(search, { code })
+      const input = await waitFor(() => query(element))
+      fireEvent.keyDown(input, { code })
 
       await waitFor(() => {
-        expect(handleSpy).toHaveBeenCalledTimes(3)
+        const userActions = userActionHistory.getUserActionsHistory()
+        expect(handleSpy).toHaveBeenCalledOnce()
+        expect(userActions[0].type).to.eq(expectedUserActionType)
       })
     }
+  })
+
+  describe('typing in an text field triggers a change event', () => {
+    const inputTestId = 'input-testid'
+
+    describe('when input is not a text field', () => {
+      const nonTextInputTypes = ['radio', 'checkbox', 'button']
+      for (const inputType of nonTextInputTypes) {
+        it(`does not trigger a Change event when type is ${inputType}`, async () => {
+          const { element, domWindow } = createElementInJSDOM(
+            `
+              <div>
+                  <input data-testid="${inputTestId}" type="${inputType}" />
+              </div>`,
+            'div',
+          )
+
+          new KeyDownEventListener(userActionHandler, domWindow, userActionHistory).init()
+          const input = await waitFor(() => getByTestId(element, inputTestId))
+          fireEvent.keyDown(input, { code: 'KeyA' })
+
+          await waitFor(() => {}, { timeout: 200 })
+          await waitFor(() => {
+            expect(userActionHistory.getUserActionsHistory().map((userAction) => userAction.type)).not.to.contain(
+              UserActionType.Change,
+            )
+          })
+        })
+      }
+    })
+
+    describe('when input is a text field', () => {
+      const fields = [
+        { type: 'text', html: `<input type="text" data-testid="${inputTestId}" />` },
+        { type: 'search', html: `<input type="search" data-testid="${inputTestId}" />` },
+        { type: 'email', html: `<input type="email" data-testid="${inputTestId}" />` },
+        { type: 'password', html: `<input type="password" data-testid="${inputTestId}" />` },
+        { type: 'textarea', html: `<textarea data-testid="${inputTestId}"></textarea>` },
+      ]
+
+      for (const { type, html } of fields) {
+        it(`does triggers a Change event when typing in ${type}`, async () => {
+          const { element, domWindow } = createElementInJSDOM(
+            `
+              <div>${html}</div>`,
+            'div',
+          )
+
+          new KeyDownEventListener(userActionHandler, domWindow, userActionHistory).init()
+          const input = await waitFor(() => getByTestId(element, inputTestId))
+          fireEvent.keyDown(input, { code: 'KeyA' })
+
+          await waitFor(() => {}, { timeout: 200 })
+          await waitFor(() => {
+            const changeEvents = userActionHistory
+              .getUserActionsHistory()
+              .filter((userAction) => userAction.type === UserActionType.Change)
+            expect(changeEvents.length).to.eq(1)
+          })
+        })
+
+        it(`does not trigger multiple Change events when typing in ${type}`, async () => {
+          const { element, domWindow } = createElementInJSDOM(
+            `
+              <div>${html}</div>`,
+            'div',
+          )
+
+          new KeyDownEventListener(userActionHandler, domWindow, userActionHistory).init()
+          const input = await waitFor(() => getByTestId(element, inputTestId))
+          fireEvent.keyDown(input, { code: 'KeyA' })
+          fireEvent.keyDown(input, { code: 'KeyB' })
+          fireEvent.keyDown(input, { code: 'KeyC' })
+          fireEvent.keyDown(input, { code: 'KeyD' })
+
+          await waitFor(() => {}, { timeout: 200 })
+          await waitFor(() => {
+            const changeEvents = userActionHistory
+              .getUserActionsHistory()
+              .filter((userAction) => userAction.type === UserActionType.Change)
+
+            expect(changeEvents.length).to.eq(1)
+          })
+        })
+
+        it('generates a value for the target', async () => {
+          const { element, domWindow } = createElementInJSDOM(
+            `
+              <div>${html}</div>`,
+            'div',
+          )
+
+          new KeyDownEventListener(userActionHandler, domWindow, userActionHistory).init()
+          const input = await waitFor(() => getByTestId(element, inputTestId))
+          fireEvent.keyDown(input, { code: 'KeyA' })
+
+          await waitFor(() => {}, { timeout: 200 })
+          await waitFor(() => {
+            const changeEvents = userActionHistory
+              .getUserActionsHistory()
+              .filter((userAction) => userAction.type === UserActionType.Change)
+            const targetedUserAction = changeEvents[0] as TargetedUserAction
+            expect(targetedUserAction.target.value).to.eq(`{{${type}}}`)
+          })
+        })
+      }
+    })
   })
 })