+++ src/event-listeners/KeyDownEventListener.ts
@@ -1,10 +1,11 @@
 import { createTargetedUserAction } from '../user-action/createTargetedUserAction'
-import { KeyUserActionData, TargetedUserAction, UserActionType } from '../types'
-import { isKeyAllowedByKeyListeners, isTargetAllowedByKeyListeners } from '../utils/listeners'
+import { HTMLInputWithValue, KeyUserActionData, TargetedUserAction, UserActionType } from '../types'
+import { isKeyAllowedByKeyListeners, isTargetAllowedByKeyListeners, recordChangeEvent } from '../utils/listeners'
 import UserActionHandler from '../user-action/UserActionHandler'
 import UserActionsHistory from '../user-actions-history/UserActionsHistory'
 import TargetedEventListener, { TargetEventListenerOptions } from './TargetedEventListener'
 import isTargetedUserAction from '../utils/isTargetedUserAction'
+import { sanitizeHTMLElementValue } from '../utils/sanitizeHTMLElementValue'
 
 class KeyDownEventListener extends TargetedEventListener {
   constructor(
@@ -20,6 +21,14 @@ class KeyDownEventListener extends TargetedEventListener {
     const userAction = createTargetedUserAction(event, this.userActionType, this.options)
     if (userAction === null || this.actionIsTheSameThanLast(userAction)) return
 
+    if (recordChangeEvent(event.code, event.target)) {
+      const changeUserAction = createTargetedUserAction(event, UserActionType.Change, this.options)
+      if (changeUserAction != null && !this.changeActionIsSame(changeUserAction)) {
+        changeUserAction.target.value = sanitizeHTMLElementValue(event.target as HTMLInputWithValue)
+        return this.userActionHandler.handle(changeUserAction)
+      }
+    }
+
     if (isKeyAllowedByKeyListeners(event.code)) {
       return this.userActionHandler.handle(userAction)
     }
@@ -29,17 +38,24 @@ class KeyDownEventListener extends TargetedEventListener {
     }
   }
 
+  changeActionIsSame(changeUserAction: TargetedUserAction): boolean {
+    const lastUserAction = this.userActionHistory.getLast()
+
+    if (lastUserAction === undefined) return false
+    if (!isTargetedUserAction(lastUserAction)) return false
+
+    return compareTargetedUserAction(changeUserAction, lastUserAction)
+  }
+
   actionIsTheSameThanLast(userAction: TargetedUserAction): boolean {
     const lastUserAction = this.userActionHistory.getLast()
+
     if (lastUserAction === undefined) return false
     if (!isTargetedUserAction(lastUserAction)) return false
     if (lastUserAction.type !== UserActionType.KeyDown) return false
+    if (!compareTargetedUserAction(lastUserAction, userAction)) return false
 
     const targetedUserAction = lastUserAction
-
-    if (targetedUserAction.target.element !== userAction.target.element) return false
-    if (targetedUserAction.target.selector !== userAction.target.selector) return false
-
     if (
       (targetedUserAction.userActionData as KeyUserActionData).key !==
       (userAction.userActionData as KeyUserActionData).key
@@ -52,6 +68,21 @@ class KeyDownEventListener extends TargetedEventListener {
       (userAction.userActionData as KeyUserActionData).code
     )
   }
+}
+
+function compareTargetedUserAction(tua1: TargetedUserAction, tua2: TargetedUserAction): boolean {
+  return sameJSONObjects(makeMinimalTargetedUserAction(tua1), makeMinimalTargetedUserAction(tua2))
+}
+
+function makeMinimalTargetedUserAction(userAction: TargetedUserAction) {
+  const { type, target } = userAction
+  const { element, selector, selectors } = target
+
+  return { type, target: { element, selector, selectors } }
+}
+
+function sameJSONObjects<T>(o1: T, o2: T): boolean {
+  return JSON.stringify(o1) === JSON.stringify(o2)
 }
 
 export default KeyDownEventListener