+++ src/event-listeners/KeyUpEventListener.test.ts
@@ -22,7 +22,7 @@ describe('KeyUpEventListener', () => {
       createTargetedUserActionSpy = vitest.spyOn(createTargetedUserActionModule, 'createTargetedUserAction')
     })
 
-    it('it calls createTargetedUserAction with the excludeRegex option', async () => {
+    it('calls createTargetedUserAction with the excludeRegex option', async () => {
       const { element, domWindow } = createElementInJSDOM(
         `
             <div>
@@ -43,7 +43,7 @@ describe('KeyUpEventListener', () => {
       })
     })
 
-    it('it calls createTargetedUserAction with the customSelector option', async () => {
+    it('calls createTargetedUserAction with the customSelector option', async () => {
       const { element, domWindow } = createElementInJSDOM(
         `
             <div>