+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased](https://github.com/Smartesting/gravity-data-collector/compare/v3.5.2...main)
+## [Unreleased](https://github.com/Smartesting/gravity-data-collector/compare/v3.7.0...main)
 
 ### Info
 
@@ -18,6 +18,30 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Removed
 
 ### Fixed
+
+## [3.7.0](https://github.com/Smartesting/gravity-data-collector/compare/v3.6.0...3.7.0)
+
+### Added
+
+- add `testingTool` property to `TestContext` to store used testing tool (e.g. Cypress)
+
+## [3.6.0](https://github.com/Smartesting/gravity-data-collector/compare/v3.5.4...v3.6.0)
+
+### Added
+
+- Store Cypress test context in tracked SessionStartedUserAction
+
+## [3.5.4](https://github.com/Smartesting/gravity-data-collector/compare/v3.5.3...v3.5.4)
+
+### Fixed
+
+- Prevent "Cannot read properties of undefined (reading 'toLowerCase')" runtime error
+
+## [3.5.3](https://github.com/Smartesting/gravity-data-collector/compare/v3.5.2...v3.5.3)
+
+### Fixed
+
+- Trigger a `Change` event when typing in a text field (`<input type="text" />` or `<textarea />`) [#20](https://github.com/Smartesting/gravity-data-collector/pull/20)
 
 ## [3.5.2](https://github.com/Smartesting/gravity-data-collector/compare/v3.5.1...v3.5.2)